Added support for basic glob patterns in "inputFile"
This allows the "inputFile" attribute of each config to include the glob
characters: *?

Any file which matches the glob pattern will be compiled.  The
"outputFile" attribute is ignored and instead the sourceFile's path is
used as the output file, but with the extension changed to the compiled
extension.  e.g. less --> css

Saving the compilerconfig.json currently does not compile all matching
files in the project.