Add ability to configure working directory for unzipping/running node modules
Added support for environment variable WEBCOMPILER_WORKING_DIRECTORY in
CompilerService.  If set, we will unzip node.exe and node modules into
a subdirectory of this directory instead of into the system temp directory.

This is intended for CI or build server environments where temp directories
are often cleaned after every run, or are set per-build-agent.